requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0

# Multimodal Support
Pillow>=10.0.0
opencv-python>=4.8.0
librosa>=0.10.0
soundfile>=0.12.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
torchvision>=0.15.0
torchaudio>=2.0.0

# Vision-Language Models
clip-interrogator>=0.6.0
open-clip-torch>=2.20.0
salesforce-lavis>=1.0.0

# Audio/Speech Models
openai-whisper>=20230314
TTS>=0.20.0
pyannote.audio>=3.0.0
speechbrain>=0.5.15

# Document Processing
pypdf>=3.15.0
python-docx>=1.0.0
pytesseract>=0.3.10
pdfplumber>=0.9.0
markdownify>=0.11.0
beautifulsoup4>=4.12.0
python-pptx>=0.6.21
openpyxl>=3.1.0
python-magic>=0.4.27
chardet>=5.2.0
ebooklib>=0.18

# Vector Stores
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support
chromadb>=0.4.15
qdrant-client>=1.6.0
weaviate-client>=3.24.0
pinecone-client>=2.2.0
redis>=5.0.0
pymongo>=4.5.0

# Language Models & NLP
langchain>=0.0.325
langsmith>=0.0.63
openai>=1.3.0
anthropic>=0.7.0
cohere>=4.37.0
tiktoken>=0.5.0
nltk>=3.8.0
spacy>=3.7.0
rank-bm25>=0.2.2

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
websockets>=12.0
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.2.0
selenium>=4.15.0
playwright>=1.40.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0
aiomysql>=0.2.0
motor>=3.3.0

# Monitoring & Logging
prometheus-client>=0.18.0
wandb>=0.15.0
mlflow>=2.8.0
tensorboard>=2.14.0
loguru>=0.7.0

# Evaluation
ragas>=0.0.22
deepeval>=0.20.0
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.4.0

# Optional GPU Optimization
# flash-attn>=2.3.0  # Requires specific CUDA compilation
# xformers>=0.0.22  # Memory efficient transformers

# Cloud Storage (Optional)
boto3>=1.28.0  # AWS S3
google-cloud-storage>=2.10.0  # Google Cloud Storage
azure-storage-blob>=12.19.0  # Azure Blob Storage